+++ HISTORY.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `useTurnstile` hook
 - `fixedSize` option to reduce layout shift
 - Missing argument for `onExpire`
+- `BoundTurnstileObject` argument to callbacks
 
 ### Fixed